+++ CHANGELOG.md
@@ -7,6 +7,12 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## [Unreleased][unreleased]
 -
 
+## [2.2.0] - 2021-01-14
+- Simplified icon hover style.
+- Updated form action icon.
+- Added support for form settings redirect success URL.
+- Removed payment data class.
+
 ## [2.1.4] - 2020-11-09
 - Improved error handling on payment start.
 - Fixed incorrect amount when using product fields.
@@ -49,7 +55,8 @@ This projects adheres to [Semantic Versioning](http://semver.org/) and [Keep a C
 ## 1.0.0 - 2015-11-05
 - First release.
 
-[unreleased]: https://github.com/wp-pay-extensions/formidable-forms/compare/2.1.4...HEAD
+[unreleased]: https://github.com/wp-pay-extensions/formidable-forms/compare/2.2.0...HEAD
+[2.2.0]: https://github.com/wp-pay-extensions/formidable-forms/compare/2.1.4...2.2.0
 [2.1.4]: https://github.com/wp-pay-extensions/formidable-forms/compare/2.1.3...2.1.4
 [2.1.3]: https://github.com/wp-pay-extensions/formidable-forms/compare/2.1.2...2.1.3
 [2.1.2]: https://github.com/wp-pay-extensions/formidable-forms/compare/2.1.1...2.1.2